requirements.txt
phidata
google-search-results
crawl4ai
groq
streamlit
python-dotenv